Lezione 13 - PubSub/pubsub_app/requirements.txt

flask
google-cloud-pubsub